# docs/swagger.yaml
basePath: /
definitions:
...
  models.Role:
    properties:
      created_on:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_on:
        type: string
    type: object
...
  models.User:
    properties:
      contact_number:
        maxLength: 20
        type: string
      created_by:
        type: string
      created_on:
        type: string
      email:
        maxLength: 255
        type: string
      expired_on:
        type: string
      first_login_on:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      last_login_on:
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      token:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - email
    - name
    - role_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
...
  /api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create user with auto-generated password (returned in response)
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - Users
  /api/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a user (soft delete)
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: user
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - Users
  /api/users/{id}/change-password:
    post:
      consumes:
      - application/json
      description: User can change their password by providing old and new password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Password change data (old_password, new_password, confirm_password)
        in: body
        name: passwordData
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - Users
  /api/users/{id}/reset-password:
    post:
      description: Admin can reset a user's password, generating a new temporary password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Reset user password (admin only)
      tags:
      - Users
  /api/users/search:
    get:
      description: Retrieve users based on provided filters (email, contact number,
        or both).
      parameters:
      - description: User Email
        in: query
        name: email
        type: string
      - description: User Contact Number
        in: query
        name: contact_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Search users by email or contact number
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"